Demos - fix build

+++ tutorials/03_parameters_enumeration/main.cpp
@@ -63,7 +63,7 @@ int main()
 			rpr_parameter_type type;
 			CHECK( rprContextGetParameterInfo(context, int(i), RPR_PARAMETER_TYPE, sizeof(type), &type, NULL));
 
-			uint64_t value_length = 0;
+			size_t value_length = 0;
 			CHECK( rprContextGetParameterInfo(context, int(i), RPR_PARAMETER_VALUE, 0, NULL, &value_length) );
 			char* paramValue = nullptr;
 			if (value_length > 0)
@@ -165,7 +165,7 @@ int main()
 			rpr_uint nodeInputType = 0;
 			CHECK(  rprMaterialNodeGetInputInfo(microfacet, i, RPR_MATERIAL_NODE_INPUT_TYPE, sizeof(nodeInputType), &nodeInputType, NULL) );
 
-			uint64_t inputValueSizeB = 0;
+			size_t inputValueSizeB = 0;
 			CHECK(  rprMaterialNodeGetInputInfo(microfacet, i, RPR_MATERIAL_NODE_INPUT_VALUE, 0, NULL, &inputValueSizeB));
 			char* materialParameterValue = new char[inputValueSizeB];
 			CHECK(  rprMaterialNodeGetInputInfo(microfacet, i, RPR_MATERIAL_NODE_INPUT_VALUE, inputValueSizeB, materialParameterValue, NULL));
